Requesting user can cancle match, requested user can reject match and fix some issues

+++ controllers/matchController.js
@@ -1,10 +1,11 @@
+const mongoose = require("mongoose");
 const Match = require("../models/match.js");
 
 const checkForExistingMatches = async (requestedTeam, date) => {
   const dateToDateString = new Date(date).toDateString();
 
   const matches = await Match.find({
-    // status: ["pending", "accepted"],
+    status: { $in: ["pending", "accepted"] },
     $or: [
       { "teams.requestingTeam.userId": requestedTeam },
       { "teams.requestedTeam.userId": requestedTeam },
@@ -14,9 +15,12 @@ const checkForExistingMatches = async (requestedTeam, date) => {
   const hasMatchValue = matches.map((item, index) => {
     let date = new Date(item.date.toDateString());
 
-    console.log(date.toDateString() === dateToDateString, "from map");
-    return date.toDateString() === dateToDateString && matches.length > 0;
+    console.log(item);
+
+    return date.toDateString() === dateToDateString && matches;
   });
+
+  console.log("match function: ", matches);
 
   return hasMatchValue;
 };
@@ -28,13 +32,6 @@ const addANewMatch = async (req, res, next) => {
     const requestingTeam = matchValues.teams.requestingTeam.userId;
     const requestedTeam = matchValues.teams.requestedTeam.userId;
 
-    console.log(
-      requestingTeam,
-      "requesting team",
-      requestedTeam,
-      "requestedTeam"
-    );
-
     const date = matchValues.date;
     const requestingTeamSquad = matchValues.squads.requestingTeamSquad.squadId;
 
@@ -45,18 +42,18 @@ const addANewMatch = async (req, res, next) => {
 
     let hasMatch = false;
 
+    console.log(requestedTeamHasMatch, "match");
+
     if (requestedTeamHasMatch.length === 0) {
       hasMatch = false;
     } else {
       hasMatch = true;
     }
 
-    console.log(requestedTeamHasMatch, "from add match");
-
     if (hasMatch) {
       const error = new Error("They have a match on this day");
       error.statusCode = 400; // Set custom status code
-      return next(error); // Pass the error to the next middleware (error handler)
+      return next(error);
     }
 
     const match = new Match({
@@ -74,7 +71,7 @@ const addANewMatch = async (req, res, next) => {
 
     res.status(201).json(match);
   } catch (e) {
-    console.log(e);
+    next(e);
   }
 };
 
@@ -114,7 +111,7 @@ const updateMatch = async (req, res, next) => {
 
     res.send(match);
   } catch (e) {
-    console.log(e);
+    next(e);
   }
 };
 
@@ -164,20 +161,59 @@ const getMatchByRequestedTeamId = async (req, res, next) => {
 
     res.status(200).json(requestedTeam);
   } catch (e) {
-    console.log(e);
+    next(e);
   }
 };
 
-const cancelMatch = async (req, res, next) => {
-  const { matchId } = req.query;
-  const match = await Match.findById(matchId);
-  if (!match) {
-    let error = new Error("There are no match found");
-    error.statusCode = 404;
+const cancelMatchByRequestingUser = async (req, res, next) => {
+  try {
+    const { matchId } = req.query;
+    const match = await Match.findById(matchId);
+    if (!match) {
+      let error = new Error("There are no match found");
+      error.statusCode = 404;
+      return next(error);
+    }
+
+    if (match.status === "pending" || match.status === "rejected") {
+      await match.deleteOne();
+      res.json({ message: "match deleted successfully" });
+    } else {
+      res.json({
+        user: match.teams.requestedTeam.userId,
+        message: "Already Accepted this match",
+      });
+    }
+  } catch (e) {
+    next(e);
+  }
+};
+
+const rejectMatchByRequestedUser = async (req, res, next) => {
+  try {
+    const { matchId, note } = req.body;
+
+    const removeRequestedUserId = new mongoose.Types.ObjectId();
+
+    console.log(matchId, "matchId", note);
+
+    const match = await Match.findById(matchId);
+    if (!match) {
+      let error = new Error("There are no Match found");
+      error.statusCode = 404;
+      return next(error);
+    }
+    console.log(match, "requestedUser");
+
+    match.teams.requestedTeam.userId = removeRequestedUserId;
+    match.status = "rejected";
+    match.note = note || "";
+
+    await match.save();
+    res.json({ message: "Rejected Match", match });
+  } catch (error) {
     return next(error);
   }
-  await match.deleteOne();
-  res.json({ message: "match deleted successfully" });
 };
 
 module.exports = {
@@ -185,5 +221,6 @@ module.exports = {
   updateMatch,
   getMatchByRequestingTeamId,
   getMatchByRequestedTeamId,
-  cancelMatch,
+  cancelMatchByRequestingUser,
+  rejectMatchByRequestedUser,
 };
